added clear-checked button and ./modules dir
