+++ 2-Webpack/src/index.html
@@ -35,6 +35,7 @@
                     <!-- to be added dynamically-->
                 </ul>
                 <button id='add-button' class='add-button'>Add a new todo item</button>
+                <button id='clean-button' class='clean-button'>Remove checked</button>
             </div>
         </div>
     </body>
